backend/requirements.txt
Django==4.2.27
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
qrcode
Pillow
reportlab
python-dotenv
sentry-sdk
django-cors-headers
gunicorn
whitenoise
dj-database-url
social-auth-app-django
